Summary Functions for table1::table1() for Numeric Data

The default functions in table1::table1() compute the Mean (SD) and Median (Min, Max) for continuous values. These summary functions are meant to offer more flexibility in terms of summaries.

Usage

table1_numeric(
  x,
  mean_sd = TRUE,
  median_iqr = TRUE,
  range = TRUE,
  quantiles = NULL,
  completeness = TRUE,
  na.rm = TRUE,
  quantile_type = 7,
  digits = 2
)

Arguments

x

A vector of values passed from table1::table1()

mean_sd

A logical scalar: Compute Mean & SD?

median_iqr

A logical scalar: Compute Median & IQR?

range

A logical scalar: Compute Range?

quantiles

A vector of numeric values: Quantiles to compute

completeness

A logical scalar: Compute completeness?

na.rm

logical scalar: Remove NA values?

quantile_type

A numeric scalar, indicating the method of computing quantiles with stats::quantile().

digits

A numeric scalar, indicating the number of significant figures to include in results.

Value

A vector of character-formatted results

Examples

table1_numeric(
  x = c(1:100, NA),
  quantiles = c(0.05, 0.95)
)
#>                     Mean (SD)   Median [IQR]     [Min, Max]        5%, 95% 
#>             "" "50.5 (29.01)" "50.5 (29.01)"     "[1, 100]"  "5.95, 95.05" 
#>  Complete (N%) 
#> "100 (99.01%)" 

data(jfbr_test)
library(table1)

  table1(
    x = ~ continuous + numbers | two_level_group,
    data = jfbr_test,
    render.continuous = table1_numeric
  )
#> <table class="Rtable1">
#> <thead>
#> <tr>
#> <th class='rowlabel firstrow lastrow'></th>
#> <th class='firstrow lastrow'><span class='stratlabel'>Group 1<br><span class='stratn'>(N=97)</span></span></th>
#> <th class='firstrow lastrow'><span class='stratlabel'>Group 2<br><span class='stratn'>(N=103)</span></span></th>
#> <th class='firstrow lastrow'><span class='stratlabel'>Overall<br><span class='stratn'>(N=200)</span></span></th>
#> </tr>
#> </thead>
#> <tbody>
#> <tr>
#> <td class='rowlabel firstrow'>continuous</td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>Mean (SD)</td>
#> <td>0.53 (0.3)</td>
#> <td>0.56 (0.29)</td>
#> <td>0.54 (0.29)</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>Median [IQR]</td>
#> <td>0.51 (0.3)</td>
#> <td>0.6 (0.29)</td>
#> <td>0.59 (0.29)</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>[Min, Max]</td>
#> <td>[0, 0.99]</td>
#> <td>[0.01, 0.99]</td>
#> <td>[0, 0.99]</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>Complete (N%)</td>
#> <td>97 (100%)</td>
#> <td>102 (99.03%)</td>
#> <td>199 (99.5%)</td>
#> </tr>
#> <tr>
#> <td class='rowlabel lastrow'>Missing</td>
#> <td class='lastrow'>0 (0%)</td>
#> <td class='lastrow'>1 (1.0%)</td>
#> <td class='lastrow'>1 (0.5%)</td>
#> </tr>
#> <tr>
#> <td class='rowlabel firstrow'>numbers</td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>Mean (SD)</td>
#> <td>100.05 (57.01)</td>
#> <td>101.42 (59.03)</td>
#> <td>100.75 (57.91)</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>Median [IQR]</td>
#> <td>100 (57.01)</td>
#> <td>103.5 (59.03)</td>
#> <td>101 (57.91)</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>[Min, Max]</td>
#> <td>[1, 200]</td>
#> <td>[2, 197]</td>
#> <td>[1, 200]</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>Complete (N%)</td>
#> <td>97 (100%)</td>
#> <td>102 (99.03%)</td>
#> <td>199 (99.5%)</td>
#> </tr>
#> <tr>
#> <td class='rowlabel lastrow'>Missing</td>
#> <td class='lastrow'>0 (0%)</td>
#> <td class='lastrow'>1 (1.0%)</td>
#> <td class='lastrow'>1 (0.5%)</td>
#> </tr>
#> </tbody>
#> </table>