Compute hypothesis tests for table1::table1()

table1::table1() has the ability to add a column for hypothesis test p-values: see ?table1::table1.

Usage

table1_pvalue(
  x,
  variable,
  test_numeric_2_levels = stats::t.test,
  test_numeric_more_than_2_levels = function(data, formula) {
     data.frame(p.value =
    stats::anova(stats::lm(formula = formula, data = data))$`Pr(>F)`[1])
 },
  test_categorical_2_levels = stats::chisq.test,
  test_categorical_more_than_2_levels = stats::chisq.test,
  digits = 3
)

Arguments

x

A named list of values - Typically passed from table1::table1().

variable

a character scalar indicating the grouping variable name - Typically passed from table1::table1().

test_numeric_2_levels

a function that computes a hypothesis test using a formula (variable ~ group) where group has only two levels, and variable is numeric. The default is t.test, alternatives include: wilcox.test.

test_numeric_more_than_2_levels

a function that computes a hypothesis test using a formula (variable ~ group) where group has more than two levels, and variable is numeric. The default is the omnibus F-test from anova(lm(variable ~ group)), alternatives include: kruskal.test.

test_categorical_2_levels

a function that computes a hypothesis test using a formula (variable ~ group) where group has only two levels and variable is categorical. The default is chisq.test, alternatives include: fisher.test.

test_categorical_more_than_2_levels

a function that computes a hypothesis test using a formula (variable ~ group) where group has more than two levels and variable is categorical. The default is chisq.test, alternatives include: fisher.test.

digits

numeric scalar: number of digits to use when formatting p-values

Value

A numeric scalar containing the p-value from the specified hypothesis tests.

Examples

set.seed(12345)
n_obs <- 200

jfbr_test <-
  data.frame(
    numbers = 1:n_obs,
    continuous = runif(n = n_obs),
    binary = rbinom(n = n_obs, size = 1, prob = 0.5),
    binary_factor =
      factor(
        x = rbinom(n = n_obs, size = 1, prob = 0.5),
        levels = 0:1,
        labels = c("0. No", "1. Yes")
      ),
    categorical = factor(sample(x = 1:4, size = n_obs, replace = TRUE)),
    ordered = ordered(sample(x = 1:4, size = n_obs, replace = TRUE))
  )

table1::table1(
  x = ~ numbers + continuous + binary + ordered | categorical,
  data = jfbr_test,
  overall = FALSE,
  extra.col =
    list("p-value" = table1_pvalue)
)
#> <table class="Rtable1">
#> <thead>
#> <tr>
#> <th class='rowlabel firstrow lastrow'></th>
#> <th class='firstrow lastrow'><span class='stratlabel'>1<br><span class='stratn'>(N=52)</span></span></th>
#> <th class='firstrow lastrow'><span class='stratlabel'>2<br><span class='stratn'>(N=45)</span></span></th>
#> <th class='firstrow lastrow'><span class='stratlabel'>3<br><span class='stratn'>(N=54)</span></span></th>
#> <th class='firstrow lastrow'><span class='stratlabel'>4<br><span class='stratn'>(N=49)</span></span></th>
#> <th class='firstrow lastrow'><span class='stratlabel'>p-value</span></th>
#> </tr>
#> </thead>
#> <tbody>
#> <tr>
#> <td class='rowlabel firstrow'>numbers</td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>Mean (SD)</td>
#> <td>98.6 (63.3)</td>
#> <td>97.1 (58.0)</td>
#> <td>104 (57.0)</td>
#> <td>102 (54.0)</td>
#> <td>0.929</td>
#> </tr>
#> <tr>
#> <td class='rowlabel lastrow'>Median [Min, Max]</td>
#> <td class='lastrow'>80.0 [3.00, 199]</td>
#> <td class='lastrow'>87.0 [12.0, 200]</td>
#> <td class='lastrow'>112 [1.00, 191]</td>
#> <td class='lastrow'>103 [9.00, 198]</td>
#> <td class='lastrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel firstrow'>continuous</td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>Mean (SD)</td>
#> <td>0.513 (0.287)</td>
#> <td>0.480 (0.302)</td>
#> <td>0.582 (0.320)</td>
#> <td>0.583 (0.260)</td>
#> <td>0.216</td>
#> </tr>
#> <tr>
#> <td class='rowlabel lastrow'>Median [Min, Max]</td>
#> <td class='lastrow'>0.510 [0.00599, 0.990]</td>
#> <td class='lastrow'>0.458 [0.00114, 0.984]</td>
#> <td class='lastrow'>0.702 [0.0334, 0.993]</td>
#> <td class='lastrow'>0.617 [0.00946, 0.974]</td>
#> <td class='lastrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel firstrow'>binary</td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>Mean (SD)</td>
#> <td>0.577 (0.499)</td>
#> <td>0.489 (0.506)</td>
#> <td>0.611 (0.492)</td>
#> <td>0.571 (0.500)</td>
#> <td>0.672</td>
#> </tr>
#> <tr>
#> <td class='rowlabel lastrow'>Median [Min, Max]</td>
#> <td class='lastrow'>1.00 [0, 1.00]</td>
#> <td class='lastrow'>0 [0, 1.00]</td>
#> <td class='lastrow'>1.00 [0, 1.00]</td>
#> <td class='lastrow'>1.00 [0, 1.00]</td>
#> <td class='lastrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel firstrow'>ordered</td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>1</td>
#> <td>14 (26.9%)</td>
#> <td>8 (17.8%)</td>
#> <td>15 (27.8%)</td>
#> <td>10 (20.4%)</td>
#> <td>0.121</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>2</td>
#> <td>11 (21.2%)</td>
#> <td>10 (22.2%)</td>
#> <td>11 (20.4%)</td>
#> <td>10 (20.4%)</td>
#> <td></td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>3</td>
#> <td>13 (25.0%)</td>
#> <td>7 (15.6%)</td>
#> <td>20 (37.0%)</td>
#> <td>16 (32.7%)</td>
#> <td></td>
#> </tr>
#> <tr>
#> <td class='rowlabel lastrow'>4</td>
#> <td class='lastrow'>14 (26.9%)</td>
#> <td class='lastrow'>20 (44.4%)</td>
#> <td class='lastrow'>8 (14.8%)</td>
#> <td class='lastrow'>13 (26.5%)</td>
#> <td class='lastrow'></td>
#> </tr>
#> </tbody>
#> </table>

table1::table1(
  x = ~ numbers + continuous + binary + ordered | binary_factor,
  data = jfbr_test,
  overall = FALSE,
  extra.col =
    list("p-value" =
           function(x, value) table1_pvalue(
             x = x,
             variable = variable,
             test_numeric_2_levels = wilcox.test,
             test_numeric_more_than_2_levels = kruskal.test,
             test_categorical_2_levels = fisher.test,
             test_categorical_more_than_2_levels = fisher.test
           )
    )
)
#> <table class="Rtable1">
#> <thead>
#> <tr>
#> <th class='rowlabel firstrow lastrow'></th>
#> <th class='firstrow lastrow'><span class='stratlabel'>0. No<br><span class='stratn'>(N=95)</span></span></th>
#> <th class='firstrow lastrow'><span class='stratlabel'>1. Yes<br><span class='stratn'>(N=105)</span></span></th>
#> <th class='firstrow lastrow'><span class='stratlabel'>p-value</span></th>
#> </tr>
#> </thead>
#> <tbody>
#> <tr>
#> <td class='rowlabel firstrow'>numbers</td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>Mean (SD)</td>
#> <td>98.0 (60.3)</td>
#> <td>103 (55.8)</td>
#> <td>0.56</td>
#> </tr>
#> <tr>
#> <td class='rowlabel lastrow'>Median [Min, Max]</td>
#> <td class='lastrow'>96.0 [1.00, 197]</td>
#> <td class='lastrow'>102 [5.00, 200]</td>
#> <td class='lastrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel firstrow'>continuous</td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>Mean (SD)</td>
#> <td>0.517 (0.307)</td>
#> <td>0.563 (0.282)</td>
#> <td>0.28</td>
#> </tr>
#> <tr>
#> <td class='rowlabel lastrow'>Median [Min, Max]</td>
#> <td class='lastrow'>0.522 [0.00114, 0.990]</td>
#> <td class='lastrow'>0.609 [0.00693, 0.993]</td>
#> <td class='lastrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel firstrow'>binary</td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>Mean (SD)</td>
#> <td>0.579 (0.496)</td>
#> <td>0.552 (0.500)</td>
#> <td>0.707</td>
#> </tr>
#> <tr>
#> <td class='rowlabel lastrow'>Median [Min, Max]</td>
#> <td class='lastrow'>1.00 [0, 1.00]</td>
#> <td class='lastrow'>1.00 [0, 1.00]</td>
#> <td class='lastrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel firstrow'>ordered</td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> <td class='firstrow'></td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>1</td>
#> <td>18 (18.9%)</td>
#> <td>29 (27.6%)</td>
#> <td>0.218</td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>2</td>
#> <td>25 (26.3%)</td>
#> <td>17 (16.2%)</td>
#> <td></td>
#> </tr>
#> <tr>
#> <td class='rowlabel'>3</td>
#> <td>28 (29.5%)</td>
#> <td>28 (26.7%)</td>
#> <td></td>
#> </tr>
#> <tr>
#> <td class='rowlabel lastrow'>4</td>
#> <td class='lastrow'>24 (25.3%)</td>
#> <td class='lastrow'>31 (29.5%)</td>
#> <td class='lastrow'></td>
#> </tr>
#> </tbody>
#> </table>