Summary Functions for table1::table1()

The default functions in table1::table1() compute the Mean (SD) and Median (Min, Max) for continuous values. These summary functions are meant to offer more flexibility in terms of summaries.

Usage

table1_numeric(
  x,
  mean_sd = TRUE,
  median_iqr = TRUE,
  range = TRUE,
  quantiles = NULL,
  na.rm = FALSE,
  quantile_type = 7,
  digits = 2
)

Arguments

x

A vector of values passed from table1::table1()

mean_sd

A logical scalar: Compute Mean & SD?

median_iqr

A logical scalar: Compute Median & IQR?

range

A logical scalar: Compute Range?

quantiles

A vector of numeric values: Quantiles to compute

na.rm

logical scalar: Remove NA values?

quantile_type

A numeric scalar, indicating the method of computing quantiles with stats::quantile().

digits

A numeric scalar, indicating the number of significant figures to include in results.

Value

A vector of character-formatted results

Examples

table1_numeric(
  x = 1:100,
  quantiles = c(0.05, 0.95)
)
#>                     Mean (SD)  Meadian [IQR]     [Min, Max]        5%, 95% 
#>             "" "50.5 (29.01)" "50.5 (29.01)"     "[1, 100]"  "5.95, 95.05"