Compute hypothesis tests for table1::table1()

table1::table1() has the ability to add a column for hypothesis test p-values: see ?table1::table1.

Usage

table1_pvalue(
  x,
  variable,
  test_numeric_2_levels = stats::t.test,
  test_numeric_more_than_2_levels = function(data, formula) {
     data.frame(p.value =
    stats::anova(stats::lm(formula = formula, data = data))$`Pr(>F)`[1])
 },
  test_categorical_2_levels = stats::chisq.test,
  test_categorical_more_than_2_levels = stats::chisq.test,
  digits = 3
)

Arguments

x

A named list of values - Typically passed from table1::table1().

variable

a character scalar indicating the grouping variable name - Typically passed from table1::table1().

test_numeric_2_levels

a function that computes a hypothesis test using a formula (variable ~ group) where group has only two levels, and variable is numeric. The default is t.test, alternatives include: wilcox.test.

test_numeric_more_than_2_levels

a function that computes a hypothesis test using a formula (variable ~ group) where group has more than two levels, and variable is numeric. The default is the omnibus F-test from anova(lm(variable ~ group)), alternatives include: kruskal.test.

test_categorical_2_levels

a function that computes a hypothesis test using a formula (variable ~ group) where group has only two levels and variable is categorical. The default is chisq.test, alternatives include: fisher.test.

test_categorical_more_than_2_levels

a function that computes a hypothesis test using a formula (variable ~ group) where group has more than two levels and variable is categorical. The default is chisq.test, alternatives include: fisher.test.

digits

numeric scalar: number of digits to use when formatting p-values

Value

A numeric scalar containing the p-value from the specified hypothesis tests.

Examples

table1::table1(
  x = ~ numbers + continuous + binary + ordered | group_two_level,
  data = jfbr_test,
  overall = FALSE,
  extra.col =
    list("p-value" = table1_pvalue)
)
#> Error in eval(predvars, data, env): object 'group_two_level' not found

table1::table1(
  x = ~ numbers + continuous + binary + ordered | group_three_level,
  data = jfbr_test,
  overall = FALSE,
  extra.col =
    list("p-value" =
           function(x, value) table1_pvalue(
             x = x,
             variable = variable,
             test_numeric_2_levels = wilcox.test,
             test_numeric_more_than_2_levels = kruskal.test,
             test_categorical_2_levels = fisher.test,
             test_categorical_more_than_2_levels = fisher.test
           )
    )
)
#> Error in eval(predvars, data, env): object 'group_three_level' not found